move things around more
