+++ domaindriven-server/domaindriven-server.cabal
@@ -26,9 +26,11 @@ source-repository head
 
 library
   exposed-modules:
+      DomainDriven
       DomainDriven.Server
       DomainDriven.Server.Config
       DomainDriven.Server.Helpers
+      DomainDriven.Server.Internal.Class
       DomainDriven.Server.Param
       DomainDriven.Server.TH
       DomainDriven.Server.Types
