+++ domaindriven/domaindriven.cabal
@@ -27,7 +27,6 @@ source-repository head
 
 library
   exposed-modules:
-      DomainDriven
       DomainDriven.Internal.Class
       DomainDriven.Internal.HasFieldName
       DomainDriven.Internal.NamedFields
@@ -107,10 +106,7 @@ test-suite domaindriven-test
   type: exitcode-stdio-1.0
   main-is: Spec.hs
   other-modules:
-      DomainDriven.Internal.NamedJsonFieldsSpec
       DomainDriven.Persistance.PostgresIORefStateSpec
-      Models.Counter
-      Models.Store
       Paths_domaindriven
   hs-source-dirs:
       test
@@ -148,8 +144,7 @@ test-suite domaindriven-test
       ViewPatterns
   ghc-options: -Wall -Wcompat -Widentities -Wincomplete-record-updates -Wincomplete-uni-patterns -Wpartial-fields -Wredundant-constraints -Wincomplete-record-updates -Wincomplete-patterns -threaded -rtsopts -with-rtsopts=-N -Wall -Wunused-packages
   build-depends:
-      QuickCheck >=2.14.2
-    , aeson >=2.0.3.0
+      aeson >=2.0.3.0
     , async >=2.2.4
     , base >=4.7 && <5
     , containers >=0.6.5.1
@@ -161,8 +156,6 @@ test-suite domaindriven-test
     , microlens >=0.4.12.0
     , openapi3 >=3.2.2
     , postgresql-simple >=0.6.4
-    , quickcheck-arbitrary-adt >=0.3.1.0
-    , quickcheck-classes >=0.6.5.0
     , servant-server >=0.19.2
     , streamly >=0.8.1.1
     , text >=1.2.5.0
